// frontend/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: 'http://localhost:5000' | (string & {});
};

export type Product = {
    id?: number;
    name?: string | null;
    description?: string | null;
    sku?: string | null;
    price?: number;
    discountPrice?: number;
    category?: string | null;
    brand?: string | null;
    stockQuantity?: number;
    rating?: number;
    reviewCount?: number;
    imageUrl?: string | null;
    createdAt?: string;
    updatedAt?: string | null;
    isAvailable?: boolean;
    tags?: Array<string> | null;
};

export type User = {
    id?: number;
    firstName?: string | null;
    lastName?: string | null;
    email?: string | null;
    phoneNumber?: string | null;
    address?: string | null;
    city?: string | null;
    state?: string | null;
    zipCode?: string | null;
    country?: string | null;
    dateOfBirth?: string;
    createdAt?: string;
    lastLoginAt?: string | null;
    isActive?: boolean;
    role?: string | null;
};

export type GetApiProductsData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/Products';
};

export type GetApiProductsResponses = {
    /**
     * Success
     */
    200: Array<Product>;
};

export type GetApiProductsResponse = GetApiProductsResponses[keyof GetApiProductsResponses];

export type PostApiProductsData = {
    body?: Product;
    path?: never;
    query?: never;
    url: '/api/Products';
};

export type PostApiProductsResponses = {
    /**
     * Success
     */
    200: Product;
};

export type PostApiProductsResponse = PostApiProductsResponses[keyof PostApiProductsResponses];

export type DeleteApiProductsByIdData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/api/Products/{id}';
};

export type DeleteApiProductsByIdResponses = {
    /**
     * Success
     */
    200: unknown;
};

export type GetApiProductsByIdData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/api/Products/{id}';
};

export type GetApiProductsByIdResponses = {
    /**
     * Success
     */
    200: Product;
};

export type GetApiProductsByIdResponse = GetApiProductsByIdResponses[keyof GetApiProductsByIdResponses];

export type PutApiProductsByIdData = {
    body?: Product;
    path: {
        id: number;
    };
    query?: never;
    url: '/api/Products/{id}';
};

export type PutApiProductsByIdResponses = {
    /**
     * Success
     */
    200: unknown;
};

export type GetApiUsersData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/Users';
};

export type GetApiUsersResponses = {
    /**
     * Success
     */
    200: Array<User>;
};

export type GetApiUsersResponse = GetApiUsersResponses[keyof GetApiUsersResponses];

export type PostApiUsersData = {
    body?: User;
    path?: never;
    query?: never;
    url: '/api/Users';
};

export type PostApiUsersResponses = {
    /**
     * Success
     */
    200: User;
};

export type PostApiUsersResponse = PostApiUsersResponses[keyof PostApiUsersResponses];

export type DeleteApiUsersByIdData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/api/Users/{id}';
};

export type DeleteApiUsersByIdResponses = {
    /**
     * Success
     */
    200: unknown;
};

export type GetApiUsersByIdData = {
    body?: never;
    path: {
        id: number;
    };
    query?: never;
    url: '/api/Users/{id}';
};

export type GetApiUsersByIdResponses = {
    /**
     * Success
     */
    200: User;
};

export type GetApiUsersByIdResponse = GetApiUsersByIdResponses[keyof GetApiUsersByIdResponses];

export type PutApiUsersByIdData = {
    body?: User;
    path: {
        id: number;
    };
    query?: never;
    url: '/api/Users/{id}';
};

export type PutApiUsersByIdResponses = {
    /**
     * Success
     */
    200: unknown;
};